refactor(calendar): view range / precision policy を features/calendar/domain へ切り出す
Calendar feature 内に `domain/` ディレクトリを新設し、Calendar 固有の仕様ルール
（view 表示期間 / 時刻精度ポリシー）を lib/ から分離する。

これで features/calendar/ の責務が以下に整理される:
- domain/ : Calendar 固有の仕様ルール（純粋）
- lib/    : 汎用計算・adapter・routing helper
- stores/ : UI state
- hooks/  : React/mutation orchestration
- components/ : UI

移動:
- features/calendar/lib/range.ts → features/calendar/domain/view-range.ts
- features/calendar/lib/precision.ts → features/calendar/domain/precision.ts
- 対応 test も domain/__tests__/ へ移動

新規:
- features/calendar/domain/index.ts: domain barrel

更新:
- features/calendar/index.ts: re-export を domain/ 経由に
- useCalendarData.ts: relative import path 更新

仕様変更なし。devtools / DB / UI への影響なし。precision の定数は production
consumer ゼロ（test のみ）だが policy 言語化として価値があるため domain に移動して
維持する（dead state 削除は別 PR の領域）。

検証:
- pnpm --filter @dayopt/product typecheck / lint / test:run (1709 pass) / build
- pnpm lint:boundaries (0 violations)
- pnpm --filter @dayopt/storybook build
