+++ apps/product/src/features/calendar/domain/__tests__/precision.test.ts
