+++ apps/product/src/features/calendar/domain/precision.ts
