+++ apps/product/src/features/calendar/domain/view-range.ts
