+++ apps/product/src/features/calendar/index.ts
@@ -63,10 +63,14 @@ export { useEntryContextActions } from './hooks/operations/useEntryContextAction
 export { useEntryOperations } from './hooks/operations/useEntryOperations';
 
 // =============================================================================
+// Domain（Calendar 固有の仕様ルール）
+// =============================================================================
+export { calculateViewDateRange } from './domain/view-range';
+
+// =============================================================================
 // Lib / Utils
 // =============================================================================
 export { formatCalendarDateParam, parseCalendarDateParam } from './lib/date-param';
-export { calculateViewDateRange } from './lib/range';
 export { isCalendarViewPath } from './lib/route-utils';
 
 // ここにないものはfeature内部専用
